feat(ui): Improves the Hero component and visual standardization
Improves the styling of the Hero component and makes it more abstract, increasing its reusability across different sections of the project.

Adjustments to Tailwind settings and `globals.css` were made to improve standardization and design consistency.

New images in SVG format were added to make the visual identity of the pages more professional and high-quality.